examples/test.cpp:  Use cstdint types

+++ examples/test.cpp
@@ -1,4 +1,5 @@
 #include <cstring>
+#include <cstdint>
 #include <iostream>
 #include <tuple>
 #include <limits>
@@ -11,17 +12,17 @@ using namespace std;
 using namespace protocol_helper;
 
 typedef tuple<
-    field<8, unsigned short>, // field 0
+    field<8, uint16_t>, // field 0
     field<1, bool>,           // field 1
-    field<6, unsigned short>, // field 2
-    field<17, unsigned int>,  // field 3
-    field<9, unsigned short>, // field 4
-    field<7, unsigned short>, // field 5
-    field<16, unsigned int>,  // field 6
+    field<6, uint16_t>, // field 2
+    field<17, uint32_t>,  // field 3
+    field<9, uint16_t>, // field 4
+    field<7, uint16_t>, // field 5
+    field<16, uint16_t>,  // field 6
     field<1, bool>,           // field 7
-    field<7, unsigned short>, // field 8
+    field<7, uint16_t>, // field 8
     field<1, bool>,           // field 9
-    field<9, unsigned short>  // field 10
+    field<9, uint16_t>  // field 10
     > test;
 
 typedef protocol<test> test_proto;
